pyproject.toml
[tool.poetry]
name = "AUTH APP"
description = "Yandex.Practicum. Sprint 6."
version = "2022.3.1"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Flask = "~2.0"
SQLAlchemy = "~1.4"
psycopg2-binary = "~2.9"
Flask-SQLAlchemy = "~2.5"
redis = "~4.2"
pydantic = { version = "~1.9", extras = ["dotenv"] }
gevent = "~21.12"
pyuwsgi = "~2.0"
flasgger = "~0.9"
Flask-Pydantic = "~0.9"
Flask-JWT-Extended = "~4.3"
backoff = "^1.11.1"
Flask-Migrate = "^3.1.0"
Authlib = "^0.15.5"
requests = "^2.27.1"
transliterate = "^1.10.2"
opentelemetry-api = "^1.10.0"
opentelemetry-sdk = "^1.10.0"
opentelemetry-exporter-jaeger = "^1.10.0"
opentelemetry-instrumentation-flask = "^0.29b1"
Flask-Limiter = { version = "^2.2.0", extras = ["redis"] }

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
coverage = "^6.3"
flake8 = "^4.0"
isort = "^5.8"
mypy = "^0.941"
pre-commit = "^2.17"
pylint = "^2.7.4"
pytest = "^7.0"
pytest-asyncio = "^0.18"
pytest-cov = "^3.0"
aiohttp = "^3.8.1"
asyncpg = "^0.25.0"
aioredis = "^2.0.1"
backoff = "^1.11.1"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | venv
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
multi_line_output = 3


[tool.mypy]
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
pretty = true

[[tool.mypy.overrides]]
module = "*.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true


[tool.pylint.messages_control]
max-line-length = 88
disable = [
    "missing-docstring",
    "too-few-public-methods",
    "too-many-arguments",
]

ignored-modules = ["pydantic", "flask_sqlalchemy"]
good-names = []

[tool.pylint.similarities]
ignore-imports = true
ignore-docstrings = true
ignore-comments = true


[tool.pytest.ini_options]
asyncio_mode= "strict"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(levelname)-5.5s [%(name)s] %(message)s"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
